+++ tabbycat/api/views.py
@@ -2,7 +2,6 @@ from django.db.models import Prefetch
 from dynamic_preferences.api.serializers import PreferenceSerializer
 from dynamic_preferences.api.viewsets import PerInstancePreferenceViewSet
 from rest_framework.generics import GenericAPIView, RetrieveUpdateAPIView
-from rest_framework.permissions import IsAdminUser
 from rest_framework.response import Response
 from rest_framework.reverse import reverse
 from rest_framework.viewsets import ModelViewSet
@@ -13,25 +12,7 @@ from tournaments.mixins import TournamentFromUrlMixin
 from tournaments.models import Tournament
 
 from . import serializers
-
-
-class TournamentAPIMixin(TournamentFromUrlMixin):
-    tournament_field = 'tournament'
-
-    def perform_create(self, serializer):
-        serializer.save(**{self.tournament_field: self.tournament})
-
-    def get_queryset(self):
-        return self.serializer_class.Meta.model.objects.filter(**{self.tournament_field: self.tournament})
-
-    def get_serializer_context(self):
-        context = super().get_serializer_context()
-        context['tournament'] = self.tournament
-        return context
-
-
-class AdministratorAPIMixin:
-    permission_classes = [IsAdminUser]
+from .mixins import AdministratorAPIMixin, TournamentAPIMixin, TournamentPublicAPIMixin
 
 
 class APIRootView(AdministratorAPIMixin, GenericAPIView):
